+++ lerobot/common/envs/wrappers.py
@@ -1,17 +1,7 @@
 import gymnasium as gym
-import os
-import math
-import pathlib
 import numpy as np
 import h5py
-from libero.libero import get_libero_path
-from libero.libero.envs import OffScreenRenderEnv
-from libero.libero import benchmark
-from gymnasium import spaces
-
-import re
 import time
-from PIL import Image
 import base64
 import cv2
 import numpy as np
@@ -19,6 +9,7 @@ from openai import OpenAI
 from vila_utils.utils.prompts import get_prompt
 from vila_utils.utils.decode import add_path_2d_to_img_alt_fast, add_mask_2d_to_img, get_path_from_answer
 from vila_utils.utils.encode import scale_path
+from lerobot.common.envs.utils import convert_to_uint8, resize_with_pad
 
 
 # Constants
@@ -28,63 +19,6 @@ DOWNSAMPLE_RESOLUTION = 256
 # PATH_MODEL_NAME_MASK = "vila_3b_oxe_no_droid_path_mask"
 PATH_MODEL_NAME = "vila_3b_oxe_sim_path"
 PATH_MODEL_NAME_MASK = "vila_3b_oxe_sim_path_mask"
-
-def convert_to_uint8(img: np.ndarray) -> np.ndarray:
-    """Converts an image to uint8 if it is a float image.
-
-    This is important for reducing the size of the image when sending it over the network.
-    """
-    if np.issubdtype(img.dtype, np.floating):
-        img = (255 * img).astype(np.uint8)
-    return img
-
-
-def _resize_with_pad_pil(image: Image.Image, height: int, width: int, method: int) -> Image.Image:
-    """Replicates tf.image.resize_with_pad for one image using PIL. Resizes an image to a target height and
-    width without distortion by padding with zeros.
-
-    Unlike the jax version, note that PIL uses [width, height, channel] ordering instead of [batch, h, w, c].
-    """
-    cur_width, cur_height = image.size
-    if cur_width == width and cur_height == height:
-        return image  # No need to resize if the image is already the correct size.
-
-    ratio = max(cur_width / width, cur_height / height)
-    resized_height = int(cur_height / ratio)
-    resized_width = int(cur_width / ratio)
-    resized_image = image.resize((resized_width, resized_height), resample=method)
-
-    zero_image = Image.new(resized_image.mode, (width, height), 0)
-    pad_height = max(0, int((height - resized_height) / 2))
-    pad_width = max(0, int((width - resized_width) / 2))
-    zero_image.paste(resized_image, (pad_width, pad_height))
-    assert zero_image.size == (width, height)
-    return zero_image
-
-
-def resize_with_pad(images: np.ndarray, height: int, width: int, method=Image.BILINEAR) -> np.ndarray:
-    """Replicates tf.image.resize_with_pad for multiple images using PIL. Resizes a batch of images to a target height.
-
-    Args:
-        images: A batch of images in [..., height, width, channel] format.
-        height: The target height of the image.
-        width: The target width of the image.
-        method: The interpolation method to use. Default is bilinear.
-
-    Returns:
-        The resized images in [..., height, width, channel].
-    """
-    # If the images are already the correct size, return them as is.
-    if images.shape[-3:-1] == (height, width):
-        return images
-
-    original_shape = images.shape
-
-    images = images.reshape(-1, *original_shape[-3:])
-    resized = np.stack(
-        [_resize_with_pad_pil(Image.fromarray(im), height, width, method=method) for im in images]
-    )
-    return resized.reshape(*original_shape[:-3], *resized.shape[-3:])
 
 
 def draw_onto_image(vlm_path_mask_output, prompt_type, img, verbose=False):
@@ -296,7 +230,6 @@ class GroundTruthPathMaskWrapper(gym.Wrapper):
         self.current_masked_images = None
         self.rng = np.random.default_rng()
 
-
     def _modify_observation(self, obs):
         """Applies path and mask drawing to a single observation."""
         if self.image_key not in obs["pixels"]:
@@ -304,10 +237,10 @@ class GroundTruthPathMaskWrapper(gym.Wrapper):
 
         img = obs["pixels"][self.image_key].copy()
         mask = self.current_mask[self.env.current_step % len(self.current_mask)]
-        #mask_points = np.stack(mask.nonzero(), axis=1)
-        #min_in, max_in = np.zeros(2), np.array(mask.shape)
-        #min_out, max_out = np.zeros(2), np.ones(2)
-        #mask_points = scale_path(mask_points, min_in=min_in, max_in=max_in, min_out=min_out, max_out=max_out)
+        # mask_points = np.stack(mask.nonzero(), axis=1)
+        # min_in, max_in = np.zeros(2), np.array(mask.shape)
+        # min_out, max_out = np.zeros(2), np.ones(2)
+        # mask_points = scale_path(mask_points, min_in=min_in, max_in=max_in, min_out=min_out, max_out=max_out)
         if self.draw_mask:
             # mask directly
             img = mask[..., None] * img
@@ -387,206 +320,3 @@ class GroundTruthPathMaskWrapper(gym.Wrapper):
         obs, reward, terminated, truncated, info = self.env.step(action)
         obs = self._modify_observation(obs)
         return obs, reward, terminated, truncated, info
-
-
-class LIBEROEnv(gym.Env):
-    LIBERO_DUMMY_ACTION = [0.0] * 6 + [-1.0]
-
-    def __init__(
-        self,
-        task_suite_name: str,
-        seed: int,
-        task_idx: int,
-        episode_idx: int,
-        resolution: int = 256,
-        libero_hdf5_dir: str = None,
-        load_gt_initial_states: bool = False,
-    ):
-        super().__init__()
-        self.LIBERO_ENV_RESOLUTION = resolution
-        self.num_steps_wait = 10
-        self._task = None
-        self.task_suite_name = task_suite_name
-        self.seed = seed
-        if task_suite_name == "libero_spatial":
-            self._max_episode_steps = 220  # longest training demo has 193 steps
-        elif task_suite_name == "libero_object":
-            self._max_episode_steps = 280  # longest training demo has 254 steps
-        elif task_suite_name == "libero_goal":
-            self._max_episode_steps = 300  # longest training demo has 270 steps
-        elif task_suite_name == "libero_10":
-            self._max_episode_steps = 520  # longest training demo has 505 steps
-        elif task_suite_name == "libero_90":
-            self._max_episode_steps = 400  # longest training demo has 373 steps
-        else:
-            raise ValueError(f"Unknown task suite: {task_suite_name}")
-        benchmark_dict = benchmark.get_benchmark_dict()
-        self._libero_task_suite = benchmark_dict[self.task_suite_name]()
-        self._libero_hdf5_dir = libero_hdf5_dir
-        self.load_gt_initial_states = load_gt_initial_states
-        self.current_step = 0
-        self.set_task_idx(task_idx)
-        self.set_episode_idx(episode_idx)
-        # load dummy env first
-        # env, _ = self._get_libero_env()
-        self.metadata = {"render_fps" : 10}
-        self.render_mode = "rgb_array"
-        self.observation_space = spaces.Dict(
-            {
-                "pixels": spaces.Dict(
-                    {
-                        "image": spaces.Box(
-                            0,
-                            255,
-                            shape=(self.LIBERO_ENV_RESOLUTION, self.LIBERO_ENV_RESOLUTION, 3),
-                            dtype=np.uint8,
-                        ),
-                        "image_wrist": spaces.Box(
-                            0,
-                            255,
-                            shape=(self.LIBERO_ENV_RESOLUTION, self.LIBERO_ENV_RESOLUTION, 3),
-                            dtype=np.uint8,
-                        ),
-                    }
-                ),
-                "agent_pos": spaces.Box(-np.inf, np.inf, shape=(8,)),
-            }
-        )
-        self.action_space = spaces.Box(-1.0, 1.0, shape=(7,))
-        self.spec = {}
-        self.spec["max_episode_steps"] = self._max_episode_steps
-
-    @property
-    def task(self):
-        return str(self._task)
-
-    @property
-    def episode_idx(self):
-        return self._episode_idx
-
-    @property
-    def num_tasks(self):
-        return self._libero_task_suite.n_tasks
-
-    def set_task_idx(self, task_idx: int):
-        self._task_idx = task_idx
-
-    def set_episode_idx(self, episode_idx: int):
-        self._episode_idx = episode_idx
-
-    def _construct_obs(self, obs):
-        flipped_agentview = obs["agentview_image"][::-1]
-        flipped_eye_in_hand = obs["robot0_eye_in_hand_image"][::-1]
-        new_obs = {}
-        pixels = {}
-        # following stupid lerobot hardcoded pixels naming...
-        pixels["image"] = convert_to_uint8(
-            resize_with_pad(flipped_agentview, self.LIBERO_ENV_RESOLUTION, self.LIBERO_ENV_RESOLUTION)
-        )
-        pixels["image_wrist"] = convert_to_uint8(
-            resize_with_pad(flipped_eye_in_hand, self.LIBERO_ENV_RESOLUTION, self.LIBERO_ENV_RESOLUTION)
-        )
-        new_obs["pixels"] = pixels
-        # following stupid lerobot hardcoded agent_pos naming...
-        new_obs["agent_pos"] = np.concatenate(
-            [
-                obs["robot0_eef_pos"],
-                LIBEROEnv._quat2axisangle(obs["robot0_eef_quat"]),
-                obs["robot0_gripper_qpos"],
-            ]
-        )
-        return new_obs
-
-    def reset(self, seed=None, **kwargs):
-        if seed is not None:
-            self.seed = seed
-        self.env, initial_states = self._get_libero_env()
-        obs = self.env.reset(**kwargs)
-
-        if self.load_gt_initial_states:
-            self.env.set_init_state(initial_states)
-        current_steps_waited = 0
-        while current_steps_waited < self.num_steps_wait:
-            obs, _, _, info = self.env.step(self.LIBERO_DUMMY_ACTION)
-            current_steps_waited += 1
-        self.current_step = 0
-        self.obs = self._construct_obs(obs)
-        return self.obs, info
-
-    def step(self, action):
-        obs, reward, terminated, info = self.env.step(action)
-        self.current_step += 1
-        truncated = self.current_step >= self._max_episode_steps
-        self.obs = self._construct_obs(obs)
-        if terminated:
-            info["is_success"] = True
-        else:
-            info["is_success"] = False
-        return self.obs, reward, terminated, truncated, info
-
-    def _load_initial_states_from_h5(self, episode_idx: int):
-        """Load initial states from HDF5 file."""
-        # get the hdf5 names
-        hdf5_names = os.listdir(self._libero_hdf5_dir)
-        task_name_underscore = self.task.replace(" ", "_")
-        for hdf5_name in hdf5_names:
-            if task_name_underscore not in hdf5_name:
-                continue
-            with h5py.File(os.path.join(self._libero_hdf5_dir, hdf5_name), "r", swmr=True) as f:
-                return f["data"][f"demo_{episode_idx}"]["states"][0]
-
-        raise ValueError(f"Could not find task name {self.task} in HDF5 files")
-
-    def _get_libero_env(self):
-        """Initializes and returns the LIBERO environment, along with the task description."""
-        task_description = self._libero_task_suite.get_task(self._task_idx).language
-        self._task = task_description
-        task_bddl_file = (
-            pathlib.Path(get_libero_path("bddl_files"))
-            / self._libero_task_suite.get_task(self._task_idx).problem_folder
-            / self._libero_task_suite.get_task(self._task_idx).bddl_file
-        )
-        env_args = {
-            "bddl_file_name": task_bddl_file,
-            "camera_heights": self.LIBERO_ENV_RESOLUTION,
-            "camera_widths": self.LIBERO_ENV_RESOLUTION,
-        }
-        env = OffScreenRenderEnv(**env_args)
-        env.seed(
-            self.seed
-        )  # IMPORTANT: seed seems to affect object positions even when using fixed initial state
-        if self.load_gt_initial_states:
-            initial_states = self._load_initial_states_from_h5(self._episode_idx)
-        else:
-            initial_states = None
-
-        return env, initial_states
-
-    @staticmethod
-    def _quat2axisangle(quat):
-        """
-        Copied from robosuite: https://github.com/ARISE-Initiative/robosuite/blob/eafb81f54ffc104f905ee48a16bb15f059176ad3/robosuite/utils/transform_utils.py#L490C1-L512C55
-        """
-        # clip quaternion
-        if quat[3] > 1.0:
-            quat[3] = 1.0
-        elif quat[3] < -1.0:
-            quat[3] = -1.0
-
-        den = np.sqrt(1.0 - quat[3] * quat[3])
-        if math.isclose(den, 0.0):
-            # This is (close to) a zero degree rotation, immediately return
-            return np.zeros(3)
-
-        return (quat[:3] * 2.0 * math.acos(quat[3])) / den
-
-    def render(self):
-        return self.obs["pixels"]["image"]
-
-
-class BRIDGEEnv(gym.Env):
-    """Wrapper for the BRIDGE client-server environment. This will act as a server that receives images and actions from the client and returns the next observation.
-
-    Args:
-        gym (_type_): _description_
-    """
